// src/main/resources/ruFacilities/src/app/model/directors.ts
import {Facilities} from "./facilities";

export class Directors {
  facilitiesDirectorsId: string;
  facilitiesDirectorsName: string;
  facilitiesDirectorsEmail: string;
  facilitiesDirectorsPhone: string;
  facilitiesDirectorsRoom: string;
  facilitiesDirectorsInfo: string;
  facilitiesDirectorsPosition: string;
  facilitiesDirectorsPositionName: string;
  facilitiesDirectorsEx1: string;
  facilities: Facilities;
}